MAINT-917 Strategy change: promote semantic index changes.
This commit changes the way we deal with the semantic index. Previously we did not promote semantic index changes but let the promotion commit update the semantic index based on the content changes. This however led to a bug where both the child and parent branches had new versions of a concept in the semantic index. With this commit semantic index changes are promoted with the rest of the content and there is no rebuilding during promotion.
The semantic index strategy for rebase remains the same - on rebase we clear out the semantic index of the branch and rebuild a delta of changes.

Co-Authored-By: Michael Chu <mch@ihtsdo.org>