MAINT-1199 Fix merge: include components in mainline which have end dates.
We made a fix in ElasticVC to the standard selection so that when using a timepoint in the past components which have since been ended are still visible. This is neccessary in the release fix branch merge function so that those components can be ended to prevent duplicate versions at the same timepoint.

Co-Authored-By: Michael Chu <mch@ihtsdo.org>