Support for multi-platform docker images
This commit provides support for multiple archtecture docker images e.g.
amd64 and arm64. To facilitate this the Spotify docker-maven-plugin has
been replaced by the Google Jib Maven Plugin.

Jib optimises the building and pushing of images to container registries
for Java applications. It does not require a Dockerfile rather the image
definition is specified in the Maven POM file.

Whilst Jib mainly supports the direct building on images to registries,
it is also possible to build an image to a local docker daemon - although
building multiple architectures is not supported via this method. To
overcome this limitation Maven profiles have been added to allow a
single platform image to be built.

The base image for the container has been updated to use Amazon Corretto,
although an Eclispe Temurin image could also have been used. The official
AdoptOpenJDK images on DockerHub are now deprecated in favour of the
Eclispe images.